+++ response/error.go
@@ -11,7 +11,7 @@ type Error struct {
 
 // ErrorResponse provides error code and message for errors to be handled specifically
 type ErrorResponse struct {
-	Code    uint16 `json:"code" xml:"code"`
+	Code    int    `json:"code" xml:"code"`
 	Message string `json:"message" xml:"message"`
 }
